% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{addTimeWindow}
\alias{addTimeWindow}
\title{addTimeWindow}
\usage{
addTimeWindow(data, dateCol, keyCol = NA, window, unit = "Days",
  dateFormat = "\%m/\%d/\%Y", includeUntil = FALSE, relativeTime = TRUE,
  sample = 100, offset = 0, ...)
}
\arguments{
\item{data}{data frame with the training data to which the time window definition should be added.}

\item{dateCol}{The column name in \code{data} that will be used.}

\item{keyCol}{An optional key for the sliding window (NA as default)}

\item{window}{The window length (numeric)}

\item{unit}{The window length unit. Should be one of: "Seconds", "Minutes", "Hours", "Days", "Years"}

\item{dateFormat}{provide date format for parsing. defaults to "\%m/\%d/\%Y " see strtptime for more examples i.e. "\%m/\%d/\%Y \%I:\%M:\%S \%p"}

\item{includeUntil}{a boolean indicator for whether the very last time point should be included in the time window search space. FALSE by default.}

\item{relativeTime}{a boolean indicator for whether the time series should be coded with absolute timestamps or relative to the last point. In this case all time stamps will be negative using the defined time unit (e.g. -10 days). TRUE by default.}

\item{sample}{allows defining the maximum number of time points to be included in each time series. Random points are sampled from the time series to reduce the time series resolution in order to better capture global trends and increase runtime performance. By default set to 100.}

\item{offset}{allows defining an additional time gap between that will be masked for the feature search. The entire time series will be shifted accordingly using the time window that was picked. 0 by default.}
}
\value{
The new data
}
\description{
Add a sliding window column to the data
}

