% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Algorithms.R
\docType{data}
\name{algorithms}
\alias{algorithms}
\title{All supported algorithms}
\format{An object of class \code{list} of length 5.}
\usage{
algorithms
}
\description{
Usage examples:
algorithms$scikit$RandomForest
algorithms$R$RXGBoost
}
\keyword{datasets}

