% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{cancelJob}
\alias{cancelJob}
\title{cancelJob
cancels a queued job by a job ID}
\usage{
cancelJob(jobId)
}
\arguments{
\item{jobId.}{The id of the job as defined by \code{\link{showJobs}}}
}
\value{
TRUE if succeeded. FALSE otherwise.
}
\description{
cancelJob
cancels a queued job by a job ID
}

