% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{classofCols}
\alias{classofCols}
\title{classofCols}
\usage{
classofCols(data)
}
\arguments{
\item{data:}{data.table to examine.}
}
\value{
a vector of names for all \code{data} columns
}
\description{
list the class of column for all columns in the input \code{data}
}
\examples{
grouped = groupBy(getData("titanic_train"), by =list("pclass","sex"))
classofCols(grouped$age)
flattenCols(grouped, "age")
classofCols(grouped)
}

