% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{clearCache}
\alias{clearCache}
\title{A function to clear the cache of a specific project}
\usage{
clearCache(projectName)
}
\arguments{
\item{projectName}{The project name to clear (e.g., "titanic").}
}
\value{
The response from the server.
}
\description{
A function to clear the cache of a specific project
}

