% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{colSizes}
\alias{colSizes}
\title{colSizes}
\usage{
colSizes(data)
}
\arguments{
\item{data}{dataframe / data.table to be examined}
}
\value{
a summary of items counts per row
}
\description{
Count the number of items in each column
}
\examples{
grouped = groupBy(getData("titanic_train"), by ="pclass")
rownames(grouped)
colSizes(grouped)
head(grouped)
grouped2 = groupBy(getData("titanic_train"), by =list("pclass","sex"))
rownames(grouped2)
colSizes(grouped2)
head(grouped2)
}

