% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{colSizesUnique}
\alias{colSizesUnique}
\title{colSizesUnique}
\usage{
colSizesUnique(data)
}
\arguments{
\item{data}{dataframe / data.table to be examined}
}
\value{
a summary of unique items counts per row
}
\description{
Count the unique number of items in each column
}
\examples{
grouped = groupBy(getData("titanic_train"), by ="pclass")
rownames(grouped)
colSizes(grouped)
head(grouped)
grouped2 = groupBy(getData("titanic_train"), by =list("pclass","sex"))
rownames(grouped2)
colSizes(grouped2)
head(grouped2)
}

