% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{colsWhiteList}
\alias{colsWhiteList}
\title{colsWhiteList}
\usage{
colsWhiteList(data, cols)
}
\arguments{
\item{data:}{dataframe / data data.table to modify.}

\item{cols:}{a list of column names to get.}
}
\value{
a dataframe / data.table with the requested columns will be returned.
}
\description{
get columns content from a data frame / data.table.
}

