% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{contextTypesList}
\alias{contextTypesList}
\title{contextTypesList}
\usage{
contextTypesList(geoSpatial = FALSE, graph = FALSE, invertedIndex = FALSE,
  lookupTables = FALSE, membershipSet = FALSE, shapeFile = FALSE,
  termsMap = FALSE, timeSeries = FALSE, timeSeriesMap = FALSE)
}
\arguments{
\item{geoSpatial.}{This context allows looking for geo spatial features. All rows are indexed based on their latitude and longitude information, and then all the objects in the}
}
\description{
contextTypesList
}

