% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{distinctBy}
\alias{distinctBy}
\title{distinctBy}
\usage{
distinctBy(data, by)
}
\arguments{
\item{data:}{dataframe or data.table to be grouped}

\item{by:}{a list of column names to distinct by}
}
\value{
a new data.table object with the distinct data
}
\description{
Keep only distinct rows in the input as specified by parameter "by". When multiple rows are encountered for a certain combination the first row will be selected.
}

