% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{emptyValuePolicyList}
\alias{emptyValuePolicyList}
\title{emptyValuePolicyList}
\usage{
emptyValuePolicyList(median = FALSE, average = FALSE, nan = FALSE,
  doubleNan = FALSE, negativeInfinity = FALSE)
}
\arguments{
\item{median}{Median value of a column}

\item{average}{Average value of a column}

\item{nan}{Not a number}

\item{negativeInfinity}{Negative infinity}
}
\description{
Defines how empty values in numeric columns in the original data are being handled.
Please note - only one of the following should be set to true
}

