% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\docType{class}
\name{Enrichment-class}
\alias{Enrichment}
\alias{Enrichment-class}
\title{SB object that encapsulates an enrichment job}
\description{
SB object that encapsulates an enrichment job
}
\section{Fields}{

\describe{
\item{\code{executionId}}{enrichment job id}
}}
\section{Methods}{

\describe{
\item{\code{currentStatus()}}{Get prediction job status - Started/Finished, number of lines processed, etc.}

\item{\code{getData(localFileName = NA_character_, runBlocking = TRUE)}}{If \code{runBlocking} is TRUE, block until the job finishes while showing processed rows counter, and return the result when available. If \code{runBlocking} is FALSE return the data if available, else return NULL}

\item{\code{initialize(executionId, totalRows = NA_integer_)}}{Initializes an Enrichment object using executionId.}
}}
\examples{
# Enrichment example
\donttest{
# Create a Session object from scratch
session = Session("project name", revision_id)
# Learn
session = learn("titanic", getData("titanic_train"), target="survived")
# Non blocking predict
enrichment = session$enrich(getData("titanic_test"), async=TRUE)
enrichment$currentStatus()
enrichment = prediction$getData()
head(data)
}
}

