% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\docType{data}
\name{Enrichment}
\alias{Enrichment}
\title{SB object that encapsulates an enrichment job}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Enrichment
}
\description{
SB object that encapsulates an enrichment job
}
\examples{
# Enrichment example
\donttest{
# Learn
session = learn("titanic", getData("titanic_train"), target="survived")
# Non blocking predict
enrichment = session$enrich(getData("titanic_test"), async=TRUE)
enrichment$currentStatus()
enrichment = prediction$getData()
head(data)
}
}
\keyword{datasets}

