% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Enrichment.R
\docType{data}
\name{EnrichmentJob}
\alias{EnrichmentJob}
\title{EnrichmentJob}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
# Enrichment example
\donttest{
# Learn
session = learn("titanic", getData("titanic_train"), target="survived")

# enrich function returns EnrichmentJob object
enrichmentJob = session$enrich(getData("titanic_test"))

# currentStatus() function can be used to show job's status
enrichmentJob$currentStatus()

# use data() function to retreive the job result as a DataFrame
data = enrichmentJob$data()
head(data)
}

# A job can be canceled using the cancel() function
\donttest{
session = learn("titanic", getData("titanic_train"), target="survived")
enrichmentJob = session$enrich(getData("titanic_test"))
enrichmentJob$cancel()
}
}
\description{
\section{SB object that encapsulates a enrichment job}

\section{\bold{Methods}}{

\subsection{\code{currentStatus()}}{
\emph{Print the current job status to console}
}

\subsection{\code{data(localFileName, runBlocking)}}{
\emph{Retreive the job result as a DataFrame. By default, if the data is not currently available, the function will block until the result is ready, and download it}
\itemize{
  \item \code{localFileName} - chose the name of the file the result will be stored in 
  \item \code{runBlocking} - if TRUE - wait for the result to be available and downloaded, and then return the result. If FALSE - return the data immediately if available, return NULL otherwise. Defaults to TRUE.  
}
}

\subsection{\code{cancel()}}{
\emph{Cancel the job}
}
}
}
\keyword{datasets}

