% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{excludeCols}
\alias{excludeCols}
\title{excludeCols}
\usage{
excludeCols(data, cols)
}
\arguments{
\item{data:}{dataframe / data data.table to modify.}

\item{cols:}{a list of column names to remove.}
}
\value{
if the input \code{data} is a dataframe than a dataframe with the excluded columns will be return. If the input \code{data} is a data.table object, NA will be returned and the input object will be modified.
}
\description{
Exclude columns from a data frame. Please notice that the function return value depends on the type of input. NA will be returned if the input is of type data.table and the input object will be modified. Otherwise a  data.frame object will be returned.
}
\examples{
data = getData("titanic_train")

## data.table case
datatable = data.table(data)
class(datatable)
colnames(datatable)
excludeCols(datatable, list("sex", "age"))
colnames(datatable)

## dataframe case
class(data)
colnames(data)
excluded = excludeCols(data, list("sex", "age"))
colnames(excluded)
}

