% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3functions.R
\name{featureSearch}
\alias{featureSearch}
\title{featureSearch}
\usage{
featureSearch(projectName = "temp", trainData, target, weightColumn = NA,
  weightByClass = FALSE, contextDatasets = NA,
  featureGenerationCtrl = featureGenerationControl(),
  reportingCtrl = reportingControl(), fileEncoding = NA, autoSave = TRUE,
  runBlocking = TRUE, maxDepth = NA, functionsWhiteList = NA,
  functionsBlackList = NA, booleanNumericFeatures = NA,
  numericEqualityFeatures = NA, useGraph = NA, useCustomGraphs = NA,
  customGraphsWhiteList = NA, customGraphsBlackList = NA,
  customFunctions = NA, crossRowFeatureSearch = NA, maxFeaturesCount = NA,
  autoColumnSubSets = NA, customColumnSubsets = NA,
  maxFeatureDuration = NA, overrideMaxFeatureDurationForExternalData = NA,
  useCachedFeatures = NA, allocatedMemoryMB = NA, maxCollectionSize = NA,
  produceFeatureClusteringReport = NA, produceReports = NA)
}
\arguments{
\item{projectName}{Optional string of the session name. Setting a session name is highly recommened. "temp" by default.}

\item{target}{String of the column name of in the training file that conatins the target of the prediction.}

\item{weightColumn}{Optional. String of the name of of one of the column that indicate a weighting that is assigned to each example. NA by default.}

\item{trainData:}{A data frame to analyze.}

\item{weightByClass:}{Adds a weight column with values inverse proportional to the frequency of the class. FALSE by default.}

\item{contextDatasets:}{Optional. A list of paths to context datasets to be added to the learning.}

\item{featureGenerationControl:}{A \code{\link{featureGenerationControl}} object with specific feature generation parameters.}

\item{reportingControl:}{A \code{\link{reportingControl}} object with specific reporting parameters.}

\item{fileEncoding:}{Optional. NA by default. Options are: "ISO-8859-1", "UTF-8", "US-ASCII".}

\item{autoSave:}{Optional. Automatically saves the generated session object to a file for future use. Good in cases where the connection between R and the server was interrupted or you would like to review previous models results. TRUE by default.}
}
\value{
Session object that encapsulate the feature search result.
}
\description{
Runs SparkBeyond feature enrichment only.
}
\examples{
#session = featureSearch("titanic", getData("titanic_train"), "survived")
}

