% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{flattenCols}
\alias{flattenCols}
\title{flattenCols}
\usage{
flattenCols(data, cols, keepLast = FALSE)
}
\arguments{
\item{data:}{data.table to flatten}

\item{cols:}{list of columns to flatten}

\item{keepLast:}{Boolean. Indicates whether to keep the first or last element in the series.}
}
\value{
nothing: operates on the input data object
}
\description{
Take the only the first element from each column defined in \code{cols} and set the column type into a vector
}
\examples{
grouped = groupBy(getData("titanic_train"), by =list("pclass","sex"))
typeof(grouped$age)
flattenCols(grouped, "age")
typeof(grouped$age) ## notice that the type of the column changed
}

