% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{getSBserverIOfolder}
\alias{getSBserverIOfolder}
\title{A function to get the SparkBeyond server IO folder.}
\usage{
getSBserverIOfolder()
}
\description{
A function to get the SparkBeyond server IO folder.
}

