% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.R
\name{getTitanicFilename}
\alias{getTitanicFilename}
\title{Auxiliary function to get Titanic train/test dataset filename.}
\usage{
getTitanicFilename(train = TRUE)
}
\arguments{
\item{Logical}{indicating whether to return train or test.}
}
\value{
String with the requested dataset filepath.
}
\description{
Auxiliary function to get Titanic train/test dataset filename.
}
\examples{
getTitanicFilename()
}

