% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{isLatestRpackage}
\alias{isLatestRpackage}
\title{A function to check if the current version is the latest one}
\usage{
isLatestRpackage()
}
\description{
A function to check if the current version is the latest one
}

