% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{isLatestVersion}
\alias{isLatestVersion}
\title{A function to verify if we are using the latest server version}
\usage{
isLatestVersion()
}
\value{
Boolean indicating TRUE if we are using the latest version otherwise FALSE.
}
\description{
A function to verify if we are using the latest server version
}

