% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{isServerAlive}
\alias{isServerAlive}
\title{A function to check whether the server is alive}
\usage{
isServerAlive()
}
\value{
The response from the server.
}
\description{
A function to check whether the server is alive
}

