% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{knowledgeControl}
\alias{knowledgeControl}
\title{knowledge Control}
\usage{
knowledgeControl(useGraph = FALSE, useOpenStreetMap = FALSE,
  useWeather = FALSE, useCustomGraphs = FALSE, customGraphsWhiteList = NA,
  customGraphsBlackList = NA, customFunctions = NA)
}
\arguments{
\item{useGraph:}{Optional. A boolean indicating whether the knowledge graph should be used. FALSE by default.}

\item{useOpenStreetMap:}{Optional. A boolean indicating whether Open Street Map data should be included in feature search. FALSE by default.}

\item{useCustomGraphs:}{A boolean indicating whether custom graphs should be used. FALSE by default.}

\item{customGraphsWhiteList:}{Optional. A list that filters which domains should be used. NA by default.}

\item{customGraphsBlackList:}{Optional. A list that filters which custom graphs should be ignored. NA by default.}

\item{customFunctions:}{Optional. A list of additional functions that should be incorporated to the feature search. NA by default.}
}
\description{
knowledge Control
}

