% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3functions.R
\name{learn.file}
\alias{learn.file}
\title{learn.file}
\usage{
learn.file(projectName = "temp", trainDataFilename, target,
  testDataFilename = NA, trainTestSplitRatio = 0.8, weightColumn = NA,
  weightByClass = FALSE, contextDatasets = NA,
  featureGenerationCtrl = featureGenerationControl(),
  modelBuildingCtrl = modelBuildingControl(),
  reportingCtrl = reportingControl(), verbose = FALSE, fileEncoding = NA,
  autoSave = TRUE, runBlocking = TRUE, ...)
}
\arguments{
\item{target}{String of the column name of in the training file that contains the target of the prediction.}

\item{projectName:}{Optional string of the session name. Setting a session name is highly recommended. "temp" by default.}

\item{trainDataFilename:}{Define the path to the training data to use.}

\item{testDataFilename:}{Optional. define a the path to the test data to use. NA by default.}

\item{trainTestSplitRatio:}{Optional. Double value in [0,1] to split the train file data in order to keep some data for test. 0.8 by default. Ignored if test filename was provided.}

\item{weightColumn:}{Optional. String of the name of of one of the column that indicate a weighting that is assigned to each example. NA by default.}

\item{weightByClass:}{Adds a weight column with values inverse proportional to the frequency of the class. FALSE by default.}

\item{contextDatasets:}{Optional. A list of paths to context datasets to be added to the learning.}

\item{featureGenerationControl:}{A \code{\link{featureGenerationControl}} object with specific feature generation parameters.}

\item{modelBuildingControl:}{A \code{\link{modelBuildingControl}} object with specific model building parameters.}

\item{reportingControl:}{A \code{\link{reportingControl}} object with specific reporting parameters.}

\item{fileEncoding:}{Optional. NA by default. Options are: "ISO-8859-1", "UTF-8", "US-ASCII".}

\item{autoSave:}{Optional. Automatically saves the generated session object to a file for future use. Good in cases where the connection between R and the server was interrupted or you would like to review previous models results. TRUE by default.}

\item{runBlocking:}{Block the R console while the session is running. FALSE by default.}
}
\value{
Session object that encapsulates the model.
}
\description{
Runs SparkBeyond feature enrichment and learning process.
}
\examples{
#session = learn("titanic", titanic_file_path, target = "survived", algorithmsWhiteList = list("RRandomForest"), runBlocking = TRUE)
}

