% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{learn}
\alias{learn}
\title{learn}
\usage{
learn(projectName, trainData, target, testData = NULL,
  contextDatasets = NULL, problemDefinition = problemDefinitionControl(),
  preProcessing = preProcessingControl(),
  featureGeneration = featureGenerationControl(),
  knowledge = knowledgeControl(), modelBuilding = modelBuildingControl(),
  reporting = reportingControl(), runBlocking = TRUE, ...)
}
\arguments{
\item{target}{String of the column name of in the training file that contains the target of the prediction.}

\item{projectName:}{Optional string of the session name. Setting a session name is highly recommended. "temp" by default.}

\item{trainData:}{train data to analyze.}

\item{testData:}{Optional. test data to validate model results. NA by default.}

\item{contextDatasets:}{Optional. A list of \code{\link{contextObject}} to be added to the learning.}

\item{problemDefinition:}{A \code{\link{problemDefinitionControl}} object with specific problem definition parameters.}

\item{preProcessing:}{A \code{\link{preProcessingControl}} object with specific preprocessing parameters.}

\item{featureGeneration:}{A \code{\link{featureGenerationControl}} object with specific feature generation parameters.}

\item{knowledge:}{A \code{\link{knowledgeControl}} object with specific external knowledge parameters.}

\item{modelBuilding:}{A \code{\link{modelBuildingControl}} object with specific model building parameters.}

\item{reporting:}{A \code{\link{reportingControl}} object with specific reporting parameters.}

\item{runBlocking:}{Block the R console while the session is running. FALSE by default.}
}
\value{
Session object that encapsulates the model.
}
\description{
Runs SparkBeyond feature enrichment and learning process.
}
\examples{
#session = learn("titanic", getData("titanic_train"), target = "survived")
}

