% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{limitTimeSeries}
\alias{limitTimeSeries}
\title{limitTimeSeries}
\usage{
limitTimeSeries(data, dateCol = "SB_times_col", fromDate = NA,
  untilDate = NA, datesFormat = "\%m/\%d/\%Y")
}
\description{
limitTimeSeries
}

