% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{offsetTime}
\alias{offsetTime}
\title{offsetTime}
\usage{
offsetTime(data, dateCol = "SB_times_col", refDate,
  datesFormat = "\%m/\%d/\%Y", units = "days")
}
\arguments{
\item{data:}{data.table to be modified.}

\item{dateCol:}{The column name in \code{data} that will be modified. "SB_times_col" by default}

\item{refDate:}{The reference date to use.}

\item{datesFormat:}{the format of the from/until dates. "\%m/\%d/\%Y" by default.}

\item{units:}{the time span unit to use when creating the reference. Based on \code{\link[base]{difftime}} definitions. "days" by default.}
}
\value{
NA will be returned and the input file will be modified.
}
\description{
Offsets a time-date column by a reference date to create a relative time series with respect to the reference date. Currently work only on data.table
}

