% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prediction.R
\docType{data}
\name{Prediction}
\alias{Prediction}
\title{SB object that encapsulates a prediction job}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
Prediction
}
\description{
SB object that encapsulates a prediction job
}
\section{Fields}{

\describe{
\item{\code{executionId}}{prediction job id}
}}
\examples{
# Prediction example
\donttest{
# Learn
session = learn("titanic", getData("titanic_train"), target="survived")
# Non blocking predict
prediction = session$predict(getData("titanic_test"), async=TRUE)
prediction$currentStatus()
data = prediction$getData()
head(data)
}
}
\keyword{datasets}

