% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Radapter.R
\name{print.evaluation}
\alias{print.evaluation}
\title{Prints just a summary of the model evaluation.}
\usage{
\method{print}{evaluation}(modelPath)
}
\arguments{
\item{modelPath}{path to the model file returned by SBlearn.}
}
\description{
Prints just a summary of the model evaluation.
}
\examples{
print.evaluation(SBlearn(..)$artifactPath)
}

