% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{printSBserverHost}
\alias{printSBserverHost}
\title{A function to print the SparkBeyond server host.}
\usage{
printSBserverHost()
}
\description{
A function to print the SparkBeyond server host.
}

