% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{printSBserverIOfolder}
\alias{printSBserverIOfolder}
\title{A function to print the SparkBeyond server IO folder.}
\usage{
printSBserverIOfolder()
}
\description{
A function to print the SparkBeyond server IO folder.
}

