% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{projectRevisions}
\alias{projectRevisions}
\title{projectRevisions}
\usage{
projectRevisions(projectName)
}
\arguments{
\item{projectName}{name of project}
}
\description{
Shows information on previous revisions of the project
}

