% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{removeProjectLock}
\alias{removeProjectLock}
\title{removeProjectLock
remove a lock by id (see \code{\link{showProjectsLocks}})}
\usage{
removeProjectLock(lockId)
}
\description{
removeProjectLock
remove a lock by id (see \code{\link{showProjectsLocks}})
}

