% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{reportsList}
\alias{reportsList}
\title{Shows explantion of all the reports produced by the SparkBeyond enginet}
\usage{
reportsList()
}
\description{
Shows explantion of all the reports produced by the SparkBeyond enginet
}

