% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.R
\name{run_SB_examples}
\alias{run_SB_examples}
\title{Run SparkBeyond titanic example.}
\usage{
run_SB_examples(configuration = "1")
}
\arguments{
\item{configuration}{Configuration ID as string. 1 - use only J48, 2 - use RRandomForest, 3 - use defaultList}
}
\description{
Run SparkBeyond titanic example.
}
\examples{
# run_SB_examples()
}

