% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.R
\name{runTitanicLearn}
\alias{runTitanicLearn}
\title{Perform learn on SparkBeyond titanic example.}
\usage{
runTitanicLearn(configuration = "1", runBlocking = TRUE)
}
\arguments{
\item{configuration}{Configuration ID as string. 1 - use only J48, 2 - use RRandomForest, 3 - use defaultList.}
}
\value{
SBmodel object containing the result of the learning
}
\description{
Perform learn on SparkBeyond titanic example.
}
\examples{
model = runTitanicLearn()
}

