% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.R
\name{runTitanicTestEnrich}
\alias{runTitanicTestEnrich}
\title{Enrich titanic test dataset with SparkBeyond features.}
\usage{
runTitanicTestEnrich(model, featureCount = 10)
}
\arguments{
\item{a}{model generated by the titanic learn example.}
}
\value{
data.frame containing regular + enriched values.
}
\description{
Enrich titanic test dataset with SparkBeyond features.
}
\examples{
enriched = runTitanicTestEnrich(model)
}

