% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.R
\name{runTitanicTestPredict}
\alias{runTitanicTestPredict}
\title{Perform prediction on titanic test dataset.}
\usage{
runTitanicTestPredict(model)
}
\arguments{
\item{a}{model generated by the titanic learn example.}
}
\value{
data.frame containing a prediction for each sample in the test set.
}
\description{
Perform prediction on titanic test dataset.
}
\examples{
runTitanicTestPredict()
}

