% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{sampleDataAbsolute}
\alias{sampleDataAbsolute}
\title{sampleData}
\usage{
sampleDataAbsolute(data, count)
}
\arguments{
\item{data}{dataframe to be sampled}

\item{count}{the number of elements to keep}
}
\value{
a sampled dataset
}
\description{
Sample data by an absolute row count
}

