% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{sampleDataByClass}
\alias{sampleDataByClass}
\title{Sample data by ratio}
\usage{
sampleDataByClass(data, columnName, labels = NA, desiredDistribution = NA)
}
\arguments{
\item{data}{dataframe to be sampled}

\item{columnName}{name of column with with labels by which to modify label distributions}

\item{labels}{Optional. labels}
}
\value{
a sampled dataset with the requested distributions
}
\description{
Sample data by ratio
}
\examples{
hist(sampleDataByClass(getData("titanic_train"), "survived", c(0,1), c(3,1))$survived, plot = FALSE)$counts
}

