% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Radapter.R
\name{SBenrich}
\alias{SBenrich}
\title{Run SparkBeyond enrichment on a result of SBlearn.}
\usage{
SBenrich(modelPath, dataPath, outputPath, featureCount = NA,
  server_port = 9000)
}
\arguments{
\item{modelPath}{path to the model file returned by SBlearn.}

\item{dataPath}{path to the data to be enriched.}

\item{outputPath}{path to write the results of the enriched data.}

\item{featureCount}{Integer value signaling how many enriched features would be returned. NA by default - marking maximum number possible (based on the globalFeatureIterations parameter in SBlearn).}

\item{server_port}{the port to be accessed in the SparkBeyond API server. 9000 by default.}
}
\value{
A data frame containing the enrichedData.
}
\description{
Run SparkBeyond enrichment on a result of SBlearn.
}
\examples{
enriched = SBenrich(modelRes$artifactPath, titanic_test_filename, "./titanic_test.tsv.gz", featureCount=10)
}

