% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Radapter.R
\name{SBlearn}
\alias{SBlearn}
\title{Run SparkBeyond feature enrichment and learning process.}
\usage{
SBlearn(sessionName, trainingFilePath, target)
}
\arguments{
\item{sessionName}{String of the session name.}

\item{trainingFilePath}{path to the file to be trained on.}

\item{target}{column name of in the training file that conatins the target of the prediction.}
}
\value{
model path of the prediction on \code{trainingFilePath}.
}
\description{
Run SparkBeyond feature enrichment and learning process.
}
\examples{
modelRes = SBlearn("titanic", titanic_train_filename, "survived")
}

