% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBmodel.R
\name{SBloadModel}
\alias{SBloadModel}
\title{load saved SparkBeyond model saved using $save}
\usage{
SBloadModel(file)
}
\arguments{
\item{file}{filename where the model was saved}
}
\value{
SBmodel object
}
\description{
load saved SparkBeyond model saved using $save
}
\examples{
#tf = paste(tempfile(), ".Rdata", sep="")
#model$save(tf)
#SBloadModel(tf)
}

