% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBmodel.R
\docType{class}
\name{SBmodel-class}
\alias{SBmodel}
\alias{SBmodel-class}
\title{SB object that encapsulates a model result}
\description{
SB object that encapsulates a model result
}
\section{Fields}{

\describe{
\item{\code{artifactLoc}}{String location pointing to the model artifact.}
}}
\section{Methods}{

\describe{
\item{\code{enrich(dataPath, outputPath, featureCount = NA)}}{Returns a data frame containing the enrichedData. \code{dataPath} is the path to the file to be tested. \code{outputPath} is the path to write the results of the prediction. featureCount Integer value signaling how many enriched features would be returned. NA by default - marking maximum number possible (based on the number of features requested in modeling).}

\item{\code{evaluate()}}{Returns an evaluation object containing various information on the run including evaluation metric that was used, evaluation score, precision, confusion matrix, number of correct and incorrect instances, AUC information and more.}

\item{\code{initialize(artifact_loc, server_port = 9000)}}{initializes a model using a string provided as \code{loc}.}

\item{\code{predict(dataPath, outputPath)}}{Returns prediction on a created model. \code{dataPath} is the path to the file to be tested. \code{outputPath} is the path to write the results of the prediction.}

\item{\code{save(file)}}{Save the current object on the file in R external object format.}

\item{\code{showConfusionMatrix(normalized = FALSE, showInIDE = TRUE)}}{Shows a confusion matrix of a model in the IDE viewer on in the web browser.}

\item{\code{showExtractors(showInIDE = TRUE)}}{Shows extractors used by a model in the IDE viewer on in the web browser.}

\item{\code{showFeatures(showInIDE = TRUE)}}{Shows features used by a model in the IDE viewer on in the web browser.}

\item{\code{showFields(showInIDE = TRUE)}}{Shows fields used by a model in the IDE viewer on in the web browser.}

\item{\code{showFunctions(showInIDE = TRUE)}}{Shows functions used by a model in the IDE viewer on in the web browser.}

\item{\code{showInputSchema(showInIDE = TRUE)}}{Shows the input schema a model in the IDE viewer on in the web browser.}

\item{\code{showModelComparison(showInIDE = TRUE)}}{Shows cross validation of various algorithms tested to create a model in the IDE viewer on in the web browser.}

\item{\code{showROC()}}{Shows ROC of the model in the IDE viewer on in the web browser.}

\item{\code{showROC_CV()}}{Shows ROC of cross validation of various algorithms tested to create a model in the IDE viewer on in the web browser.}

\item{\code{showReport(report_name = NA, showInIDE = TRUE)}}{\code{report_name} should be one of the following: confusionMatrix, confusionMatrix_normalized, extractor, features, field, function, InputSchema, modelComparison, roc_best, roc_CV}
}}
\examples{
#model learn
model = SBlearn("titanic", getTitanicFilename(train = TRUE), "survived",algorithmsWhiteList = list("RRandomForest"))
#model = SBfeatureSearchOnly("titanic", getTitanicFilename(train = TRUE), "survived")
enriched = model$enrich(getTitanicFilename(train = FALSE), paste(getwd(),"titanic_test_enriched.tsv.gz",sep="/"), featureCount = 10)
colnames(enriched)
predicted = model$predict(getTitanicFilename(train = FALSE), paste(getwd(),"titanic_test_predicted.tsv.gz",sep="/"))
colnames(predicted)
predicted[1:5,c("survived_predicted", "X_0_probability", "X_1_probability")]
eval = model$evaluate()
#model$showFeatures()
#model$showConfusionMatrix()
}

