% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Radapter.R
\name{SBmodelEvaluation}
\alias{SBmodelEvaluation}
\title{Returns an SB evaluation object that contains info on a model result.}
\usage{
SBmodelEvaluation(modelPath)
}
\arguments{
\item{modelPath}{path to the model file returned by SBlearn}
}
\value{
An evaluation object containing various information on the run including evaluation metric that was used, evaluation score, precision, confusion matrix, number of correct and incorrect instances, AUC information and more.
}
\description{
Returns an SB evaluation object that contains info on a model result.
}
\examples{
evaluation = SBmodelEvaluation(SBlearn(..)$artifactPath)
}

