% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{serverVersion}
\alias{serverVersion}
\title{A function to get the server version information}
\usage{
serverVersion()
}
\value{
The server version information.
}
\description{
A function to get the server version information
}

