% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Session.R
\docType{class}
\name{Session-class}
\alias{Session}
\alias{Session-class}
\title{SB object that encapsulates a session}
\description{
SB object that encapsulates a session
}
\section{Fields}{

\describe{
\item{\code{artifact_loc}}{String location pointing to the model artifact.}

\item{\code{modelBuilt}}{Indication for whether only a feature search was performed or a full model was created.}
}}
\section{Methods}{

\describe{
\item{\code{buildNumber()}}{Returns the build number in which the model was generated.}

\item{\code{createPackage(sampleData = NULL)}}{Create a sharable package for the model. \code{sampleData} can be used to a sample data to the package and test it. Only first 20 rows of the sample data will be used.}

\item{\code{enrich(data, featureCount = NA, writePredictionColumnsOnly = TRUE)}}{Returns a data frame containing the enrichedData. \code{data} is a dataframe to be enriched. Set \code{featureCount} in order to limit the number of returned features. Set \code{writePredictionColumnsOnly} to TRUE to return only prediction and probabily columns rather than the entire dataset.}

\item{\code{enrich.file(file, featureCount = NA, writePredictionColumnsOnly = TRUE)}}{Returns a data frame containing the enrichedData. \code{file} is a path to the file to be enriched. Set \code{featureCount} in order to limit the number of returned features. Set \code{writePredictionColumnsOnly} to TRUE to return only prediction and probabily columns rather than the entire dataset.}

\item{\code{evaluate()}}{Returns an evaluation object containing various information on the run including evaluation metric that was used, evaluation score, precision, confusion matrix, number of correct and incorrect instances, AUC information and more.}

\item{\code{features()}}{Returns a dataset with top feature information}

\item{\code{initialize(artifact_loc, modelBuilt = FALSE)}}{initializes a session using a string provided as \code{loc}.}

\item{\code{liftFromPrediction(predictionResult, overrideDesiredClass = NA, title = NA,
  percentOfPopulationToPlot = 0.2)}}{Returns lift from a created model and generates three plots. \code{predictionResult} is a dataframe to be analyzed, \code{overrideDesiredClass} the class in the label column to check the lift for (e.g. '1'), \code{title} optional: a title for the plot. \code{percentOfPopulationToPlot} optional: limit the plot to the top percent of the data (x axis).}

\item{\code{predict(data, writePredictionColumnsOnly = TRUE)}}{Returns prediction on a created model. \code{data} is a dataframe to be predicted. Set \code{writePredictionColumnsOnly} to TRUE to return only prediction and probabily columns rather than the entire dataset.}

\item{\code{predict.file(file, writePredictionColumnsOnly = TRUE)}}{Returns prediction on a created model. \code{file} is the path of the file to be predicted. Set \code{writePredictionColumnsOnly} to TRUE to return only prediction and probabily columns rather than the entire dataset.}

\item{\code{showConfusionMatrix(normalized = FALSE)}}{Shows a confusion matrix of a model.}

\item{\code{showExtractors()}}{Shows extractors.}

\item{\code{showFeatureClusters()}}{Shows the representative feature clusters pdf report.}

\item{\code{showFeaturesTest()}}{Shows features performance on test.}

\item{\code{showFeaturesTrain()}}{Shows features performance on train.}

\item{\code{showFields()}}{Shows fields.}

\item{\code{showFunctions()}}{Shows functions.}

\item{\code{showInputSchema()}}{Shows the input schema.}

\item{\code{showModelComparison()}}{Shows cross validation of various algorithms tested to create a model.}

\item{\code{showROC()}}{Shows ROC of the model.}

\item{\code{showROC_CV()}}{Shows ROC of cross validation of various algorithms tested to create a model.}

\item{\code{showReport(report_name = NA)}}{\code{report_name} should be one of the following: confusionMatrix, confusionMatrix_normalized, extractor, train_features, train_unweighted_features, test_unweighted_features,field, function, InputSchema, modelComparison, roc_best, roc_CV}

\item{\code{status()}}{Checking the status of the session.}

\item{\code{waitForProcess()}}{Blocking the R console until session is finished.}
}}
\examples{
# Learning example
\donttest {
session = learn("titanic", getData("titanic_train"), target="survived",algorithmsWhiteList = list("xgboostClassifier"),  scoreOnTestSet = TRUE, useCachedFeatures=TRUE)
#Do Not Run: session = featureSearch("titanic", getData("titanic_train"), target="survived")
enriched = session$enrich(getData("titanic_train"), featureCount = 10)
colnames(enriched)
predicted = session$predict(getData("titanic_test"))
colnames(predicted)
predicted[1:5,c("survived_predicted", "probability_0", "probability_1")]
eval = session$evaluate()
#session$showFeatures()
#session$showConfusionMatrix()
}
}

