% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBmodel.R
\name{setSBserverHost}
\alias{setSBserverHost}
\title{A function to set the SparkBeyond server host.}
\usage{
setSBserverHost(host = "http://127.0.0.1")
}
\arguments{
\item{host}{new host URL.}
}
\description{
A function to set the SparkBeyond server host.
}

