% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{setSBserverIOfolder}
\alias{setSBserverIOfolder}
\title{A function to set the SparkBeyond server I/O folder.}
\usage{
setSBserverIOfolder(folder)
}
\arguments{
\item{port}{new port.}
}
\description{
A function to set the SparkBeyond server I/O folder.
}

