% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{setSBserverPort}
\alias{setSBserverPort}
\title{A function to set the SparkBeyond server port.}
\usage{
setSBserverPort(port = "9000")
}
\arguments{
\item{port}{new port.}
}
\description{
A function to set the SparkBeyond server port.
}

