% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{showJobById}
\alias{showJobById}
\title{showJobById
get information for a specific job by job ID}
\usage{
showJobById(jobId)
}
\arguments{
\item{jobId.}{The id of the job as defined by \code{\link{showJobs}}}
}
\description{
showJobById
get information for a specific job by job ID
}

