% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{showJobs}
\alias{showJobs}
\title{showJobs
shows the status for all jobs}
\usage{
showJobs(projectName = NA, revision = NA, status = NA,
  showAllColumns = FALSE)
}
\arguments{
\item{projectName}{Filter by project name.}

\item{revision}{Option to filter the returned list by revision number}

\item{status}{Filter by status.}

\item{showAllColumns}{A switch for whether to show only the job ID, project name, status, and elapsed (if available). Alternatively show all columns}
}
\value{
a data frame with the jobs
}
\description{
A status of a job can be one of "queued", "running", "failed", "canceled", "done"
}

