% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{showProjectsLocks}
\alias{showProjectsLocks}
\title{showProjectsLocks
Shows all the active projects that acquired a lock to prevent multiple project run under the same name}
\usage{
showProjectsLocks()
}
\description{
showProjectsLocks
Shows all the active projects that acquired a lock to prevent multiple project run under the same name
}

