% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{timeWindowControl}
\alias{timeWindowControl}
\title{timeWindowControl}
\usage{
timeWindowControl(dateCol, keyCol = NULL, window, unit = "Days",
  relativeTime = TRUE, sample = 100, offset = 0,
  allowHistoricalStats = TRUE, naturalModality = "No_Modality",
  applyModalityFilter = FALSE)
}
\arguments{
\item{dateCol}{The column name in training set that will be used.}

\item{keyCol}{An optional key for the sliding window (NULL as default).}

\item{window}{The window length (numeric)}

\item{unit}{The window length unit. Should be one of: "Seconds", "Minutes", "Hours", "Days", "Weeks", "Years"}

\item{relativeTime}{a boolean indicator for whether the time series should be coded with absolute timestamps or relative to the last point. In this case all time stamps will be negative using the defined time unit (e.g. -10 days). TRUE by default.}

\item{sample}{allows defining the maximum number of time points to be included in each time series. Random points are sampled from the time series to reduce the time series resolution in order to better capture global trends and increase runtime performance. By default set to 100.}

\item{offset}{allows defining an additional time gap between that will be masked for the feature search. The entire time series will be shifted accordingly using the time window that was picked. 0 by default.}

\item{allowHistoricalStats}{when true, statistics for various historical quantities of the corresponding key will be calculated over the entire history (e.g. the average historical price of an item since the beginning of time), as opposed to just within the time-window (e.g. the average price for an item within the past 3 days). TRUE by default."}

\item{naturalModality}{usefull if there are natural time cycle dependencies such as daily or monthly. Multi_Modal will cause the system to try them all and No_Modality to try none. Possible values: No_Modality, Hour_Of_Day, Day_Of_Week, Month_Of_Year, Multi_Modal. No_Modality by default."}

\item{applyModalityFilter}{a boolean indicator for whether to use only dates with the same modality as the date column, or search on both the whole window and the modal version. FALSE by default."}
}
\value{
timeWindowControl object
}
\description{
timeWindowControl
}

