% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{trainTestSplit}
\usage{
trainTestSplit(data, ratio = 0.8, seed = 1)
}
\arguments{
\item{data}{dataframe to be splitted}

\item{seed}{- Optional. A single numeric value representing the seed for the random sampling. It is guaranteed that for the same data and the same seed, the same sampling will be generated.}
}
\value{
a list of two elements with train dataframe as first element, and test dataframe as second element
}
\description{
Split data to train + test datasets by ratio
}

