% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{typeofCols}
\alias{typeofCols}
\title{typeofCols}
\usage{
typeofCols(data)
}
\arguments{
\item{data:}{data.table to examine.}
}
\value{
a vector of names for all \code{data} columns
}
\description{
list the type of column for all columns in the input \code{data}
}
\examples{
grouped = groupBy(getData("titanic_train"), by =list("pclass","sex"))
typeofCols(grouped$age)
flattenCols(grouped, "age")
typeofCols(grouped)
}

