% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{uploadFile}
\alias{uploadFile}
\title{uploadFile}
\usage{
uploadFile(filePath, projectName, resourceName, checkIfExists = TRUE)
}
\arguments{
\item{filePath}{relative or absolute path to the file that should be upoloaded to the server}

\item{projectName}{the name of the project to save the data under}

\item{resourceName}{resource name on server}

\item{checkIfExists}{check if the file was already uploaded to server}
}
\description{
uploadFile
}

