% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{uploadFileToServer}
\alias{uploadFileToServer}
\title{uploadFileToServer}
\usage{
uploadFileToServer(filePath, projectName, name = NA, generateHash = TRUE)
}
\arguments{
\item{filePath}{relative or absolute path to the file that should be upoloaded to the server}

\item{projectName}{the name of the project to save the data under}

\item{name}{the prefix of the file name in which the data will be saved}
}
\description{
uploadFileToServer
}

