% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{uploadToServer}
\alias{uploadToServer}
\title{uploadToServer}
\usage{
uploadToServer(data, projectName, name, useEscaping = TRUE,
  directUploadThreshold = NA)
}
\arguments{
\item{data}{dataFrame to be written to the server}

\item{projectName}{the name of the project to save the data under}

\item{name}{the prefix of the file name in which the data will be saved}

\item{useEscaping}{A binary indicator noting whether a forward slash in the data needs to be escaped}
}
\description{
uploadToServer
}

