% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3functions.R
\name{writeDataToServer}
\alias{writeDataToServer}
\title{A function to write a dateframe to the server. Useful for passing a dataframe to the server for feature search / learning purposes.}
\usage{
writeDataToServer(data)
}
\arguments{
\item{data}{Data frame or table to export to the server.}
}
\value{
A filepath to the file on the server that was created.
}
\description{
A function to write a dateframe to the server. Useful for passing a dataframe to the server for feature search / learning purposes.
}

