% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{writeGroupedData}
\alias{writeGroupedData}
\title{sugar to convert all columns to text and write to file}
\usage{
writeGroupedData(data, outputFile, groupColumns = NULL)
}
\description{
sugar to convert all columns to text and write to file
}

