% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{writeGroupedDataExplicit}
\alias{writeGroupedDataExplicit}
\title{sugar to convert all columns to text and write to file}
\usage{
writeGroupedDataExplicit(data, outputFile, groupCols)
}
\arguments{
\item{groupByColumns}{Optional. A vector of possible columns that were used for grouping the data. NULL by default.}
}
\description{
sugar to convert all columns to text and write to file
}

