% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{writeToFile}
\alias{writeToFile}
\title{writeToFile}
\usage{
writeToFile(data, outputFile)
}
\arguments{
\item{data:}{dataframe / data data.table to write.}
}
\description{
Convert all \code{data} columns into a serializable text/primitive representation and write them to a tab separated file.
}
\examples{
grouped = groupBy(getData("titanic_train"), by =list("pclass","sex"))
writeToFile(grouped, "titanic_grouped.tsv")
}

