% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{writeToServer}
\alias{writeToServer}
\title{writeToServer}
\usage{
writeToServer(data, filename = NA, prefix = "data_in", useEscaping = TRUE)
}
\arguments{
\item{data}{Data frame or table to export to the server.}

\item{filename}{Optional. define a name to save the data to. NA by default.}

\item{useEscaping}{A binary indicator noting whether a forward slash in the data needs to be escaped}
}
\value{
A filepath to the file on the server that was created.
}
\description{
A function to write a dateframe to the server. Useful for passing a dataframe to the server for feature search / learning purposes.
}

