% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{addTimeWindow}
\alias{addTimeWindow}
\title{addTimeWindow}
\usage{
addTimeWindow(data, dateCol, keyCol = NA, window, unit = "Days",
  dateFormat = "\%m/\%d/\%Y", includeUntil = FALSE, relativeTime = TRUE,
  sample = 2147483647, offset = 0, ...)
}
\arguments{
\item{data:}{data.table to be modified.}

\item{dateCol:}{The column name in \code{data} that will be used.}

\item{window:}{The window length (numeric)}

\item{unit:}{The window length unit. Should be one of: "Seconds", "Minutes", "Hours", "Days", "Weeks", "Years", "Number"}

\item{dateFormat:}{provide date format for parsing. defaults to "\%m/\%d/\%Y " see strtptime for more examples i.e. "\%m/\%d/\%Y \%I:\%M:\%S \%p"}

\item{keyCol:}{An optional key for the sliding window (NA as default)}

\item{includeUntil:}{optional argument}

\item{relativeTime:}{optional relative time boolean flag}

\item{offset:}{optional offset to target. 0 by default.}

\item{sample:}{optional maximal possible sample value (default to maxint)}
}
\value{
The new data
}
\description{
Add a sliding window column to the data
}

