% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{cols2Text}
\alias{cols2Text}
\title{cols2Text}
\usage{
cols2Text(data, useEscaping = TRUE)
}
\arguments{
\item{data}{dataframe / data.table to convert to primitive represenation.}

\item{useEscaping}{A binary indicator noting whether a forward slash in the data needs to be escaped}
}
\value{
a vector containing all input information in text/primitive representation
}
\description{
Convert all \code{data} columns into a serializable text/primitive representation. Cells that contain lists will be converted to [ , , ] representation. All strings will quoted.
}
\examples{
grouped = groupBy(getData("titanic_train"), by =list("pclass","sex"))
textGrouped = cols2Text(grouped)
typeofCols(textGrouped)
textGrouped[1,]
}

