% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3functions.R
\name{contextObject}
\alias{contextObject}
\title{contextObject}
\usage{
contextObject(data, name = NULL, keyColumns = list(), timeColumn = NULL)
}
\arguments{
\item{data:}{a data frame or a path to a file containing the context data}

\item{name:}{an identifier for the context object to be created (optional).}

\item{keyColumns:}{Specify the key columns to be used by the context object (optional).}

\item{timeColumn:}{Specify the time column to be used by the context object (relevant in time series contexts) (optional).}
}
\description{
contextObject
}

