% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{contextTypesList}
\alias{contextTypesList}
\title{contextTypesList}
\usage{
contextTypesList(geoSpatial = FALSE, graph = FALSE, invertedIndex = FALSE,
  lookupTables = FALSE, membershipSet = FALSE, shapeFile = FALSE,
  termsMap = FALSE, timeSeries = FALSE, timeSeriesMap = FALSE)
}
\arguments{
\item{geoSpatial.}{This context allows looking for geo spatial features. All rows are indexed based on their coordinate information, allowing properties from the surrounding environment to be searched for in a KNN-fashion. A coordinate column is required for this context object to be created.}

\item{graph.}{This context allows looking for features on a graph or a network. This context requires defining the edges in the graph by setting the source column and target column in the contextObject.}

\item{invertedIndex.}{}
}
\description{
contextTypesList
}

