% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{functionCatalog}
\alias{functionCatalog}
\title{functionCatalog
Shows SparkBeyond current function catalog with possible}
\usage{
functionCatalog()
}
\description{
functionCatalog
Shows SparkBeyond current function catalog with possible
}

