% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{getData}
\alias{getData}
\title{Auxiliary function to get public test datasets.}
\usage{
getData(datasetName)
}
\arguments{
\item{String}{of the name of the dataset. Options are:
\itemize{
\item 1. titanic_train
\item 2. titanic_test
\item 3. tweets_sentiment
\item 4. tweets_sentimentkaggle
\item 5. sentiment_lexicon
}}
}
\value{
DataFrame containing the data.
}
\description{
Auxiliary function to get public test datasets.
}
\examples{
# Do Not Run:
# getData("titanic_train")
# getData("titanic_test")
}

