% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{getData}
\alias{getData}
\title{Auxiliary function to get public test datasets.}
\usage{
getData(datasetName)
}
\arguments{
\item{String}{of the name of the dataset. Should be one of "titanic_train", "titanic_test", "flights_delay"}
}
\value{
DataFrame containing the data.
}
\description{
Auxiliary function to get public test datasets.
}
\examples{
# Do Not Run:
# getData("titanic_train")
# getData("titanic_test")
}

