% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{knowledgeControl}
\alias{knowledgeControl}
\title{knowledge Control}
\usage{
knowledgeControl(linkedDataCore = FALSE, openStreetMap = FALSE,
  weather = FALSE, usCensus = FALSE, news = FALSE, twitter = FALSE,
  searchEngines = FALSE, customDatasets = FALSE, ...)
}
\arguments{
\item{linkedDataCore}{Includes DBPedia, Yago2, Wordnet and OpenLibrary. DBpedia is a crowd-sourced community effort to extract structured information from Wikipedia. Matches training data by Text elements.}

\item{openStreetMap}{OpenStreetMap (OSM) is a collaborative project to create a free editable map of the world. Matches training data by location (coordinate).}

\item{weather}{Weather data from NOAA Climate.gov. Matches the training data by date and location.}

\item{usCensus}{Experimental: The United States Census decennial census. Matches training data by Zipcode.}

\item{news}{Experimental: News data from the GDELT Project that monitors the world's broadcast, print, and web news. Matches the training data based on a date, location or both.}

\item{twitter}{Experimental: information from tweets that appeared on a certain date (based on dates in the data). (Limited by the maximum allowed Twitter quota per day).}

\item{searchEngines}{Experimental: Use search engines to find matching datasets. Matches training data by textual elements}

\item{customDatasets}{Custom datasets current include include US zipcode information with 44 parameters, UK postcode information, data on the largest 1000 cities and airport information.}
}
\description{
knowledge Control
}

