% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SBsettings.R
\name{login}
\alias{login}
\title{login
Login to SparkBeyond server}
\usage{
login(username, password, domain)
}
\arguments{
\item{username}{Username (usually in email format)}

\item{password}{Password}

\item{domain}{Domain name or ip of the SparkBeyond Server. (Usually starts with http or https. May require also the port of the server).}
}
\description{
login
Login to SparkBeyond server
}

