% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{preProcessingControl}
\alias{preProcessingControl}
\title{preProcessingContorl}
\usage{
preProcessingControl(emptyValuePolicy = NA, fileEncoding = NA)
}
\arguments{
\item{emptyValuePolicy:}{Controls how empty values in the data are being handled NA by default will replace numeric value with median value, and strings with empty strings.}

\item{fileEncoding:}{Optional. Options are: "ISO-8859-1", "UTF-8", "US-ASCII". NA by default will try to automatically find the best encoding.}
}
\description{
preProcessingContorl
}

