% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{preProcessingControl}
\alias{preProcessingControl}
\title{preProcessingContorl}
\usage{
preProcessingControl(trainTestSplitRatio = 0.8, temporalSplitColumn = NA,
  emptyValuePolicy = NA, fileEncoding = NA, fileEscaping = TRUE)
}
\arguments{
\item{trainTestSplitRatio:}{Optional. Double value in [0,1] to split the train file data in order to keep some data for test. 0.8 by default. Ignored if test filename was provided.}

\item{emptyValuePolicy:}{Controls how empty values in the data are being handled NA by default will replace numeric value with median value, and strings with empty strings.}

\item{fileEncoding:}{Optional. Options are: "ISO-8859-1", "UTF-8", "US-ASCII". NA by default will try to automatically find the best encoding.}

\item{fileEscaping:}{Define how escaping (e.g., \n) should be handled when files are parsed.}
}
\description{
preProcessingContorl
}

