% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S3functions.R
\name{preProcessingControl}
\alias{preProcessingControl}
\title{preProcessingControl}
\usage{
preProcessingControl(fileEscaping = TRUE, fileEncoding = NA,
  linesForTypeDetection = NA, emptyValuePolicy = NA)
}
\arguments{
\item{fileEscaping}{Define how escaping (e.g., \\n) should be handled when files are parsed.}

\item{fileEncoding}{Optional. Options are: "ISO-8859-1", "UTF-8", "US-ASCII". NA by default will try to automatically find the best encoding.}

\item{linesForTypeDetection}{maximum number of lines that should be used for type detection. 10000 by default.}

\item{emptyValuePolicy}{Controls how empty values in numeric columns in the input data are being handled. DoubleNan is default see \code{\link{emptyValuePolicyList}}.}
}
\description{
preProcessingControl
}

