% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3functions.R
\name{reportingControl}
\alias{reportingControl}
\title{reportingContorl}
\usage{
reportingControl(produceFeatureClusteringReport = FALSE,
  produceReports = NA, scoreOnTestSet = FALSE)
}
\arguments{
\item{produceFeatureClusteringReport:}{An indicator to produce feature cluster visualization. FALSE by default.}

\item{produceReports:}{"EVALUATED_FUNCTIONS".}

\item{scoreOnTestSet:}{Optional. A boolean representing whether scoring should be provided for the test set. FALSE by default.}
}
\description{
reportingContorl
}

