% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{run_SB_examples}
\alias{run_SB_examples}
\title{Run SparkBeyond titanic example.}
\usage{
run_SB_examples(configuration = "1", weightByClass = FALSE)
}
\arguments{
\item{configuration}{Configuration ID as string. 1 - use RRandomForest, 2 - use defaultList}
}
\description{
Run SparkBeyond titanic example.
}
\examples{
# run_SB_examples()
}

