% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.R
\name{run_SB_examples}
\alias{run_SB_examples}
\title{Run SparkBeyond titanic example.}
\usage{
run_SB_examples(configuration = 1, server_port = 9000)
}
\arguments{
\item{server_port}{the port to be accessed in the SparkBeyond API server. 9000 by default}
}
\description{
Run SparkBeyond titanic example.
}
\examples{
run_SB_examples()
}

