% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.R
\name{runTitanicFeatureSelectionOnly}
\alias{runTitanicFeatureSelectionOnly}
\title{Perform feature search only on Titanic train data.}
\usage{
runTitanicFeatureSelectionOnly(server_port = 9000)
}
\arguments{
\item{server_port}{the port to be accessed in the SparkBeyond API server. 9000 by default.}
}
\value{
data.frame containing a prediction for each sample in the test set.
}
\description{
Perform feature search only on Titanic train data.
}
\examples{
model = runTitanicFeatureSelectionOnly()
}

