% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/example.R
\name{runTitanicLearn}
\alias{runTitanicLearn}
\title{Perform learn on SparkBeyond titanic example.}
\usage{
runTitanicLearn(configuration = "1", server_port = 9000)
}
\arguments{
\item{configuration}{Configuration ID as string. 1 - use only J48, 2 - use RRandomForest, 3 - use defaultList.}

\item{server_port}{the port to be accessed in the SparkBeyond API server. 9000 by default.}
}
\value{
SBmodel object containing the result of the learning
}
\description{
Perform learn on SparkBeyond titanic example.
}
\examples{
runTitanicLearn()
}

