% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{sampleData}
\alias{sampleData}
\title{sampleData}
\usage{
sampleData(data, ratio = 0.8)
}
\arguments{
\item{data}{dataframe to be sampled}
}
\value{
a sampled dataset
}
\description{
Sample data by ratio
}

