% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{sampleDataAbsolute}
\alias{sampleDataAbsolute}
\title{sampleData}
\usage{
sampleDataAbsolute(data, count, seed = 1)
}
\arguments{
\item{data}{dataframe to be sampled}

\item{count}{the number of elements to keep}

\item{seed}{- Optional. A single numeric value representing the seed for the random sampling. It is guaranteed that for the same data and the same seed, the same sampling will be generated.}
}
\value{
a sampled dataset
}
\description{
Sample data by an absolute row count
}

