% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{sampleDataByClass}
\alias{sampleDataByClass}
\title{sampleDataByClass}
\usage{
sampleDataByClass(data, columnName, labels = NA, desiredDistribution = NA,
  seed = 1)
}
\arguments{
\item{data}{dataframe to be sampled}

\item{columnName}{name of column with with labels by which to modify label distributions}

\item{labels}{Optional. labels in columnName to which \code{desiredDistribution} will relate to. Default NA will assume equal distributions between all labels.}

\item{desiredDistribution}{Optional. The requested distribution ratio of \code{labels}. Default NA will assume equal distributions between all labels.}

\item{seed}{- Optional. A single numeric value representing the seed for the random sampling. It is guaranteed that for the same data and the same seed, the same sampling will be generated.}
}
\value{
a sampled dataset with the requested distributions
}
\description{
Under/over sample data based on labels of \code{columnName}.
}
\examples{
hist(sampleDataByClass(getData("titanic_train"), "survived", c(0,1), c(3,1))$survived, plot = FALSE)$counts
}

