% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/S3functions.R
\name{SBfeatureSearchOnly}
\alias{SBfeatureSearchOnly}
\title{Run SparkBeyond feature enrichment and learning process.}
\usage{
SBfeatureSearchOnly(projectName = "temp", trainData, trainDataFilename = "",
  overridePreviousFile = TRUE, target, weightColumn = NA, maxDepth = 2,
  hints = NA, useGraph = FALSE, maxFeaturesCount = 300,
  allocatedMemoryMB = 1000, runBlocking = FALSE)
}
\arguments{
\item{projectName}{Optional string of the session name. Setting a session name is highly recommened. "temp" by default.}

\item{target}{String of the column name of in the training file that conatins the target of the prediction.}

\item{weightColumn}{Optional. String of the name of of one of the column that indicate a weighting that is assigned to each example. NA by default.}

\item{maxDepth}{Optional. Integer < 8 which represent the maximun number of transformations allowed during the feature search phase. Increasing this value should be considered with cautious as the feature search phase is exponential. 2 by default.}

\item{hints}{Optional. A list of strings that reprents a set of hints that will be used to guide the feature search. NA by default.}

\item{useGraph}{Optional. A boolean indicating whether the knowledge graph should be used. FALSE by default.}

\item{maxFeaturesCount}{Optional. An integer of how many features should be created by the SB engine. 300 by default.}

\item{trainingFilePath}{String of the path to the file to be trained on.}
}
\value{
SBmodel object that encapsulate the feature search result.
}
\description{
Run SparkBeyond feature enrichment and learning process.
}
\examples{
#model = SBfeatureSearchOnly("titanic", titanic_train_filename, "survived")
}

