% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/SBmodel.R
\docType{class}
\name{SBmodel-class}
\alias{SBmodel}
\alias{SBmodel-class}
\title{SB object that encapsulates a model result}
\description{
SB object that encapsulates a model result
}
\section{Fields}{

\describe{
\item{\code{artifact_loc}}{String location pointing to the model artifact.}

\item{\code{modelBuilt}}{Indication for whether only a feature search was performed or a full model was created.}
}}
\section{Methods}{

\describe{
\item{\code{enrich(data, featureCount = NA, dataFilename = "",
  overridePreviousFile = TRUE)}}{Returns a data frame containing the enrichedData. \code{data} is a dataframe to be tested. \code{dataFilename} is optional name to write on the server. \code{overridePreviousFile} is indicator whether to override previous file that may exist on the server. \code{featureCount} Integer value signaling how many enriched features would be returned. NA by default - marking maximum number possible (based on the number of features requested in modeling).}

\item{\code{evaluate()}}{Returns an evaluation object containing various information on the run including evaluation metric that was used, evaluation score, precision, confusion matrix, number of correct and incorrect instances, AUC information and more.}

\item{\code{features()}}{Returns a dataset with top feature information}

\item{\code{initialize(artifact_loc, modelBuilt = FALSE)}}{initializes a model using a string provided as \code{loc}.}

\item{\code{predict(data, dataFilename = "", overridePreviousFile = TRUE)}}{Returns prediction on a created model. \code{data} is a dataframe to be tested. \code{dataFilename} is optional name to write on the server. \code{overridePreviousFile} is indicator whether to override previous file that may exist on the server.}

\item{\code{showConfusionMatrix(normalized = FALSE, showInIDE = TRUE)}}{Shows a confusion matrix of a model in the IDE viewer on in the web browser.}

\item{\code{showExtractors(showInIDE = TRUE)}}{Shows extractors used by a model in the IDE viewer on in the web browser.}

\item{\code{showFeatures(showInIDE = TRUE)}}{Shows features used by a model in the IDE viewer on in the web browser.}

\item{\code{showFields(showInIDE = TRUE)}}{Shows fields used by a model in the IDE viewer on in the web browser.}

\item{\code{showFunctions(showInIDE = TRUE)}}{Shows functions used by a model in the IDE viewer on in the web browser.}

\item{\code{showInputSchema(showInIDE = TRUE)}}{Shows the input schema a model in the IDE viewer on in the web browser.}

\item{\code{showModelComparison(showInIDE = TRUE)}}{Shows cross validation of various algorithms tested to create a model in the IDE viewer on in the web browser.}

\item{\code{showROC(showInIDE = TRUE)}}{Shows ROC of the model in the IDE viewer on in the web browser.}

\item{\code{showROC_CV(showInIDE = TRUE)}}{Shows ROC of cross validation of various algorithms tested to create a model in the IDE viewer on in the web browser.}

\item{\code{showReport(report_name = NA, showInIDE = TRUE)}}{\code{report_name} should be one of the following: confusionMatrix, confusionMatrix_normalized, extractor, features, field, function, InputSchema, modelComparison, roc_best, roc_CV}
}}
\examples{
#model learn
model = SBlearn("titanic", getTitanicData(train = TRUE), "survived",algorithmsWhiteList = list("RRandomForest"))
#model = SBfeatureSearchOnly("titanic", getTitanicData(train = TRUE), "survived")
enriched = model$enrich(getTitanicData(train = FALSE), featureCount = 10)
colnames(enriched)
predicted = model$predict(getTitanicData(train = FALSE))
colnames(predicted)
predicted[1:5,c("survived_predicted", "X_0_probability", "X_1_probability")]
eval = model$evaluate()
#model$showFeatures()
#model$showConfusionMatrix()
}

