% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/Radapter.R
\name{SBpredict}
\alias{SBpredict}
\title{Run SparkBeyond prediction on a result of SBlearn.}
\usage{
SBpredict(modelPath, dataPath, outputPath)
}
\arguments{
\item{modelPath}{path to the model file returned by SBlearn.}

\item{dataPath}{path to the file to be tested.}

\item{outputPath}{path to write the results of the prediction.}
}
\value{
A data frame containing the prediction.
}
\description{
Run SparkBeyond prediction on a result of SBlearn.
}
\examples{
predictRes = SBpredict(modelRes$artifactPath, titanic_test_filename, "./titanic_test.tsv.gz")
}

