% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/helperFunctions.R
\name{trainTestSplit}
\alias{trainTestSplit}
\title{Split data to train + test datasets by ratio}
\usage{
trainTestSplit(data, ratio = 0.8)
}
\arguments{
\item{data}{dataframe to be splitted}
}
\value{
a list of two elements with train dataframe as first element, and test dataframe as second element
}
\description{
Split data to train + test datasets by ratio
}

