% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complexTypes.R
\name{writeToFile}
\alias{writeToFile}
\title{writeToFile}
\usage{
writeToFile(data, outputFile, useEscaping = TRUE)
}
\arguments{
\item{data}{dataframe / data data.table to write.}

\item{useEscaping}{A binary indicator noting whether a forward slash in the data needs to be escaped}
}
\description{
Convert all \code{data} columns into a serializable text/primitive representation and write them to a tab separated file.
}
\examples{
grouped = groupBy(getData("titanic_train"), by =list("pclass","sex"))
writeToFile(grouped, "titanic_grouped.tsv")
}

